requirements.txt
flask
upstash-redis
python-dotenv
